end Sum and Mul

+++ main/header.h
@@ -1,7 +1,7 @@
 #ifndef HEADER
 #define HEADER
-double MySum(int, int);
-double MyMul(int, int);
+double MySum(double, double);
+double MyMul(double, double);
 double MySub(double, double);
 double MyDiv(double, double);
